Replace entity listeners with custom events dispatched from Doctrine onFlush event
This makes the logic working more consistently, without need to
differentiate limitations and quirks of each Doctrine lifecycle event.

Thanks to this change last update date in the User and RadioTable entity
will not be refreshed just because user clicked "save changes" button
without actually changing anything in the form.